401 error

+++ src/main/java/com/workintech/s19_twitter_challange/controller/CommentController.java
@@ -3,6 +3,7 @@ package com.workintech.s19_twitter_challange.controller;
 import com.workintech.s19_twitter_challange.dto.CommentRequestDto;
 import com.workintech.s19_twitter_challange.dto.CommentResponseDto;
 import com.workintech.s19_twitter_challange.entity.User;
+import com.workintech.s19_twitter_challange.exceptions.UserNotLoginAlreadyException;
 import com.workintech.s19_twitter_challange.service.CommentService;
 import jakarta.validation.constraints.Positive;
 import lombok.AllArgsConstructor;
@@ -22,25 +23,32 @@ public class CommentController {
     @PostMapping("/{tweetId}")
     @ResponseStatus(HttpStatus.CREATED)
     public CommentResponseDto save(@AuthenticationPrincipal User user, @Positive @PathVariable long tweetId, @Validated @RequestBody CommentRequestDto comment){
-
+        if (user == null) {
+            throw new UserNotLoginAlreadyException("Giriş yapmalısınız!");
+        }else{
             long userId = user.getId();
             return commentService.save(userId,tweetId,comment);
+        }
     }
 
     @PatchMapping("/{tweetId}/{commentId}")
     public CommentResponseDto update(@AuthenticationPrincipal User user,@Positive @PathVariable("tweetId") long tweetId,
                           @Positive @PathVariable("commentId") long commentId,@Validated @RequestBody CommentRequestDto comment){
-
+        if (user == null) {
+            throw new UserNotLoginAlreadyException("Giriş yapmalısınız!");
+        }else{
             long userId = user.getId();
             return commentService.update(userId,tweetId,commentId,comment);
-
+        }
     }
     @DeleteMapping("/{tweetId}/{commentId}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public CommentResponseDto delete(@AuthenticationPrincipal User user,@Positive @PathVariable("tweetId") long tweetId,@Positive @PathVariable("commentId") long commentId){
-
+        if (user == null) {
+            throw new UserNotLoginAlreadyException("Giriş yapmalısınız!");
+        }else{
             long userId = user.getId();
             return  commentService.delete(userId,tweetId,commentId);
-
+        }
     }
 }
